Started code to add/remove users from clubs

+++ src/scripts/removeclubuser.php
@@ -1,8 +1,10 @@
 <?php
 session_start();
 
-$_SESSION["error"] = "db couldn't connect";
-include("db_connect_test.php");
-$sql = "DELETE FROM ClubMember WHERE user_id = {$_POST['user_id']} AND club_id = {$_SESSION['curr_club']}";
-$_SESSION["error"] = $sql;
-$db->query($sql);
+if (isset($_SESSION["acc_type"]) && isset($_SESSION["curr_club"])) {
+    if ($_SESSION["acc_type"] == "admin" || $_SESSION["acc_type"] == "club_admin") {
+        include("db_connect_test.php");
+        $sql = "DELETE FROM ClubMember WHERE user_id = {$_POST['user_id']} AND club_id = {$_SESSION['curr_club']}";
+        $db->query($sql);
+    }
+}
